refactor code to create local storage instance
